api/requirements.txt
fastapi==0.111.0
uvicorn==0.30.0
kafka-python==2.0.2
pydantic==2.7.1